Commencement of custom Login Page

+++ src/main/java/com/TheTrio/SATracker/repository/UserRepository.java
@@ -1,9 +1,10 @@
 package com.TheTrio.SATracker.repository;
 
-import com.TheTrio.SATracker.models.*;
+import com.TheTrio.SATracker.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+// Use Long as the ID type because User.id is a Long
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 }
